Default to showing only entries from the current year
Specifically, the context for `books` is now initialized to a context
for the current year when the config file is created. This remains
static, so once the new year arrives it will continue showing books from
the now-previous year until you explicitly advance it via the
as-yet-unwritten `new-year` command.

This default context limits it to:
- finished books that were finished within the year
- active books that were started before the end of the year
- all inactive books

And then individual reports like `next` or `list` can filter it down
further.

tasklib then disables this context if year:foo filters are applied.